Move status updating to legal hold service
