+++ zmessaging/src/main/scala/com/waz/service/ZMessaging.scala
@@ -274,8 +274,6 @@ class ZMessaging(val teamId: Option[TeamId], val clientId: ClientId, account: Ac
   lazy val legalHold: LegalHoldService                = if (BuildConfig.LEGAL_HOLD_ENABLED) wire[LegalHoldServiceImpl]
                                                         else wire[DisabledLegalHoldService]
   lazy val legalHoldSync: LegalHoldSyncHandler        = wire[LegalHoldSyncHandlerImpl]
-  lazy val legalHoldStatusUpdater: LegalHoldStatusUpdater = if (BuildConfig.LEGAL_HOLD_ENABLED) wire[LegalHoldStatusUpdaterImpl]
-                                                            else new DummyLegalHoldStatusUpdater()
 
   lazy val eventPipeline: EventPipeline = new EventPipelineImpl(Vector(), eventScheduler.enqueue)
 
@@ -354,7 +352,6 @@ class ZMessaging(val teamId: Option[TeamId], val clientId: ClientId, account: Ac
     recordAndPlay
     messagesIndexStorage
     verificationUpdater
-    legalHoldStatusUpdater
     propertiesService
 
     rolesService.ensureDefaultRoles()
